Move the test for the urlencode_htmlentities() Utils method to test-utils.php
